search_date/requirements.txt
Flask==1.1.1
pytesseract==0.3.0
Pillow==6.2.0
datefinder==0.7.0
google-cloud-vision==0.40.0